Basic infrastructure in place for test clients.
